Begin stack to use DecoratorAssetsDefinitionBuilder for asset-check-related decorators  (#22402)
## Summary & Motivation

The general approach here will be to incrementally "eat" the current implementation of these decorators by creating the builders and then using their interior methods to incrementally replace the entire body of decorator-based construction.

In this case we will replace the call to `compute_required_resource_keys` with the one in the builder.

## How I Tested These Changes

BK