fix compute kind tag (#22504)
## Summary & Motivation

#22359 caused the op snapshots created on the code server and loaded in
a host process to have different snapshot IDs, due to patching in of the
new compute kind tag in the host process. This causes a snapshot ID
mismatch error in plus. This PR:

- Adds a backcompat test to ensure the snapshot ID for an op doesn't
change.
- Removes the snap layer host process patch in favor of a GQL layer
patch for ops.
- Patches UI to look for either the legacy or new tag when looking at
asset tags. This is necessary because these tags are not patched at the
GQL layer (they should be in a followup, and we can remove all reference
to the old tag in the UI)

## How I Tested These Changes

New test