Refactor asset backfill retry behavior to cover more classes of failures
Summary:
This replaces the retry logic in asset backfills from being user-code-server-faiure-specific to cover more classes of failures. It now works more like AMP retry logic, where the sequence of events is;

a) generate a candidate set of runs to launch, reserve run IDs for them, update the backfill data with those run IDs
b) launch each run one by one, at each step making sure that the run hasn't been launched
c) once you're done, write that you're finished and are ready to move on

This guarantees idempotence because we're resilient to failures at any stage in the above:
- If we fail before the write in a), no writes have been done, so we're good
- If we fail partway through a), we have persisted which runs we want to launch, so we

This also should let us remove some kinda gnarly existing retry logic that was specific to user code failures and had to kind of backtrack to try to recover (instead of bubbling up the failure and relying on idempotency in the next tick).