Fix observation handling for freshness checks (#22486)
When last update timestamp metadata key isn't provided for an
observation, prefer the latest materialization.

This has the potential to be wrong occasionally. The second-latest event
could actually be an observation with the correct metadata, instead of
the materialization, but figuring that out would require us performing
an expensive query.