stop rebuilding execution plan from snapshot in execute_run (#22498)
This was originally put here as part of a project to be able to execute
runs from a host process which we ended up moving away from. The only
real utilty this serves is keeping that capability excercised on a
critical path. Unfortunately, this can be the source of a class of
desync issues when the execution plan is created on version N and then
the run starts on version N+1. We also always build the execution plan
from scratch in execute_step, so this puts the run code path back in
line with that.

The changes that were required to make this work mostly fall out of the
rebuild_from_snapshot code path effectively acting as a way to pass
through `step_keys_to_execute` so we now just store that resolved value
on `DagsterRun` where previously we would record `None` to mean "execute
everything".

## How I Tested These Changes

existing coverage