[rfc] Add earliest step failure event to job failure event (#22439)
This will allow us to retrieve step failure information in a performant
way for large runs. I chose to retrieve "earliest" step failure event,
since that's the one relevant to the state transition from "in progress"
to "failed" run.

Some concern for if, in the future, we want to include more than just
the latest step failure, we'll need to have some serdes logic to switch
this new field to a list that includes more than just the latest event.
It's also possible we can choose some non-comprehensive N of step
failure events to occur, but I think I prefer just retrieving the
earliest for now to keep it simple.

Added a test to ensure that we actually receive the earliest step
failure.