Checkpoint requested_subset before re-raising when there is an error in the middle of an asset backfill
Summary:
Followup from https://github.com/dagster-io/dagster/pull/22538 - which expands our retry behavior, but loses a nice property where if there was an error in the middle of a 25 run chunk, we would still correctly update the list of requested partitions (of course this wouldn't happen if there was a hard crash, but in a more 'expected' error like a user code server being down or an agent being unavailable, it's nice to keep the list of requested partitions up to date)

This PR brings back that behavior by simplifying the chunking behavior / removing submit_asset_runs_in_chunks entirely - instead, we check hte number of runs that have been submitted and update the written backfill every 25 runs. If there's an error, we do that write as well before re-raising the exception so that the retry logic can take over.

Test Plan: BK, launch a backfill with 100 runs and shut down the code server halfway through, observe that the backfill pauses and the correct set of requested partitions are shown