Lazily import from dateutil.tz to work around default EMR python environment issue
Summary:
Workaround for the issue identified in https://github.com/dagster-io/dagster/issues/22586 where EMR has two different versions of dateutil installed in the default python environment - generally in an EMR context we are just running an op and won't need this import to resolve correctly anyway.

Also add a pin for the dateutil version that we actually depend on now.

Test Plan: BK