Add dateutil as vendored module
Summary:
dateutil is in an odd situation where it is imported by two different published pypi packages - py-dateutil and python-dateutil (and in fact the default AWS EMR python environment appears to include both), causing confusion and errors. It is also relatively small, self-contained, and updated infrequently, making it a good candidate for vendoring.

The only changes we needed to make to the source code were to turn some absolute imports into relative imports.