Bugfix handle string source key prefix (#22654)
## Summary & Motivation
To my understanding, it should be possible to provide the
`source_key_prefix` argument to `load_assets_from_modules` as a string
and not only as a list of strings, just as is the case for `key_prefix`.
However, this was not the case. Although passing validation checks,
loading a SourceAsset with key `source_asset_key` with
`source_key_prefix="prefix"` resulted in `AssetKey(['p', 'r', 'e', 'f',
'i', 'x', 'source_asset_key'])`.
Fixed the behavior by checking that `source_asset_key` is a list before
applying it.

## How I Tested These Changes
Added an additional assert to function `test_source_key_prefix` in
`test_assets_from_modules.py`.