[docs] - Add type hint to Backfill example (#22680)
## Summary & Motivation

This PR adds a type hint to an example in the Backfills documentation.
This was brought up by a user who couldn't use the example as-is without
adding `-> None` or `-> MaterializeResult`:

> a quick question - I am setting the backfill policy for an asset to be
single_run - I get an error that the IOManager does not support
partitioned assets with single run - but I am not returning anyhting,
any ideas why dagster thinks I am using an IOManager? I see in the logs
an output steps with output of "any" type
> Ok I managed to make it work by giving my function a -> None type
hint, it works also with a -> MaterializeResult type hint to only return
metadata. the example here in the documentation does not work for me:
https://docs.dagster.io/concepts/partitions-schedules-sensors/backfills#launching-single-run-backfills-using-backfill-policies
as they show they are able to trigger the run without type hints

## How I Tested These Changes