consolidate graph asset op handle dependency logic (#22669)
## Summary & Motivation

In graph-backed assets, there ops that directly produce assets often
depend on ops that don't directly produce assets. In various situations,
we need to find all the ops within a graph-backed asset that need to be
executed to materialize a particular asset. Or the inverse.

This PR aims to simplify the way we track this correspondence by
consolidating it inside `AssetsDefinition`. Prior to this PR, it's
tracked with a big dictionary in the asset layer, which is a little
confusing because it makes it seems like these dependencies can cross
`AssetsDefinition`.

The motivation for this PR is to have a smooth base for simplifications
to this logic, which will ultimately help with refactoring the
relationship between `AssetLayer` and `AssetGraph`.

Direct effects of this PR:
- `AssetLayer` loses the `dependency_node_handles_by_asset_key` and
`dep_asset_keys_by_node_output_handle` properties
- The `AssetLayer. downstream_dep_assets` method delegates to the
underlying `AssetsDefinition`
- `asset_or_check_key_to_dep_node_handles` becomes a method of
`AssetsDefinition`, and operates on a single `AssetsDefinition` instead
of a dictionary of them

## How I Tested These Changes