[useThrottledMemo] Don't Access factory via ref. (#22690)
## Summary & Motivation

Feedback on my last PR suggested I used a ref to access `factory` but I
realize now that it's wrong because it causes different semantics from
`useMemo` which uses stale closures until dependencies change. So I'm
reverting that part.

ie. In this case we would want to keep the stale closure:
```
useMemo(() => {
  arg3(arg1, arg2);
  // @ts-ignore intentionally don't include arg3 in memo for xyz reason.
}, [arg1, arg2])
```

## How I Tested These Changes
jest