[ui] Turn off ligatures in certain monospace usage (#22698)
## Summary & Motivation

Disable ligatures in Geist Mono in a number of places throughout the
app, namely in markdown output and in the config editor. (Basically
anywhere that we may show code.)

<img width="232" alt="Screenshot 2024-06-25 at 10 54 19"
src="https://github.com/dagster-io/dagster/assets/2823852/923cc3cb-b66d-4a58-97a9-d567a4bd4c62">

<img width="163" alt="Screenshot 2024-06-25 at 10 33 32"
src="https://github.com/dagster-io/dagster/assets/2823852/dde930a3-fc63-4bab-b484-1cad98ba63f1">


## How I Tested These Changes

Add a markdown description to an asset with "->" in it, verify that it
is rendered without the ligature on the asset detail page.

Add a "->" to a job launchpad editor, verify that it is not rendered
with a ligature.