fix unexpected type in JobExecutionResult._get_value() (#22722)
## Summary & Motivation

The `input_name` param of `StepExecutionContext.for_input_manager` has a
`str` type annotation. However when I made a change elsewhere that
enforced this type deeper in the stack, I found a callsite where a
`None` value is passed in for this parameter.

I considered adding a constant for the value that's being passed in, but
it's not referenced elsewhere. Can still do so though if preferred.

## How I Tested These Changes

- Added type annotations to `_get_value`
- Observed that the type checker started to complain about the value
passed to the `input_name` argument in `for_input_manager`
- Added my "fix"
- Observed that the type checker stopped complaining