don't include outer node handle in AssetsDefinition asset/op deps (#22747)
## Summary & Motivation

The `asset_or_check_keys_by_dep_op_output_handle` and
`dep_op_handles_by_asset_or_check_key` properties on `AssetsDefinition`
return dictionaries that relate asset keys to ops within the
`AssetsDefinition`.

We use `NodeHandle`s to identify the ops within the `AssetsDefinition`,
but the way that we do so is non-standard way:
- A `NodeHandle` represents a path down the composition tree to a node
within a graph, from the root of that graph. They do _not_ include the
name of the outer graph definition, because that name might be aliased
if it's embedded in a larger graph.
- The `NodeHandle`s returned in these dictionaries _do_ contain the name
of the root graph definition.

We in fact do embed these asset-level graph definitions within a larger
graph - the asset-job level graph. This means that op handles within
that job-level graph can point to the wrong op within that graph.

## How I Tested These Changes

Added an assert to `test_graph_backed_asset_reused`, which fails before
this PR and passes with it.