Fix navigating to folders in asset catalog  (#22764)
## Summary & Motivation

https://github.com/dagster-io/dagster/pull/22589/ broke the behavior for
navigating to folders within the asset catalog. Previously on an
`AssetNotFoundError` we would assume the path was a folder and render
the asset catalog with that path set, after that PR we would instead
always redirect to the root of the catalog. After this PR we instead
redirect to the catalog at that current path

## How I Tested These Changes

tests locally with app proxy