Fix warnings (#22750)
## Summary & Motivation

Limit nesting of manipulation of the `_warnings_on` context variable to
one level. The motivation here is that, in certain niche cases (in
particular observed in pytest), nested `finally` resets of the token
sometimes occur out-of-order, which leads to the `_warnings_on` context
variable being "permanently" reset to `False`. By avoiding this nesting,
we can avoid this bug (whether it's a problem in python, pytest, or our
own code, which is unclear).

This also re-enables the skipped assertions that surfaced this bug
before.

## How I Tested These Changes

Pushed this 5 times and did not observe previous inconsistently failing
tests to fail.