fix(dbt): use correct relation identifier when creating `BaseRelation` (#22823)
## Summary & Motivation
When creating the `BaseRelation`, we want to use the name of the table
that's created in the data warehouse so that we can properly query the
relation.

Previously, we were using `dbt_resource_props["name"]` as the
identifier, which is the dbt's reference to that table. Instead, we
should use the underlying identifier name.

For dbt sources, this is `dbt_resource_props["identifier"]`. For dbt
models and seeds, this is `dbt_resource_props["alias"]`.

## How I Tested These Changes
pytest: added dbt aliases for source and model in existing
`test_dagster_metadata` project