[ui] Provide an explicit Refresh button to update op partition status FE-411 (#22689)
## Summary & Motivation

A user reported internal inconsistencies on the op job partitions page
here:
https://linear.app/dagster-labs/issue/FE-411/reporting-an-issue-with-the-backfill-launcher-ui-not-consistently

There are a few things going on:

- This page never updates the overall partition status bar after page
load
- This page DOES update the run+step statuses, but only in the selected
range
- This page DOES update the backfill status shown at the very bottom of
the page in the table

This means that if a backfill is running, it's easy to get this page
into a state where it's reporting different information in different
places, especially the top bar. This is bad because it doesn't even
refresh when you click "Launch backfill", and causes the UI elements in
the modal (eg: failed checkbox) to behave incorrectly.

My guess is that we originally periodically refreshed the overall
partition status bar and took it out for performance reasons. We just
invested more into perf + caching in #22629 due to backend load, so I
think the best solution for now is a tried-and-true explicit refresh
button, and triggering a refresh when you open the launchpad.


![image](https://github.com/dagster-io/dagster/assets/1037212/480bc6f7-9cc1-4cf7-bba6-efdf66fcbac6)


## How I Tested These Changes

Tested this by launching a long-running backfill and playing with the
page extensively while it was running.

Co-authored-by: bengotow <bgotow@elementl.com>