[bugfix] Show actual creation timestamp instead of `start_time` in runs table (#22788)
## Summary & Motivation

A recent user inquiry about the sorting of runs in the backfill runs
table
(https://dagsterlabs.slack.com/archives/C064933ANQL/p1719328567183329)
led to the discovery that what is listed as the "Create date" in the
runs table UI is not actually the "Create time" of the underlying run
record, but rather the `start_time` (when the run was launched) if it
was defined, or if not the `update_time`.

This is both inaccurate and also causes the sorting in the table to be
inscrutable in some cases, where there is a subtle deviation from
sorting by "Created date".

This PR updates the UI to use the run record `creation_time` instead of
`start_time` or `updated_time`. `creation_time` is always defined since
it is set on record creation. This change both fixes the inaccuracy on
the runs page and results in more straightforward sorting.

Note that creation time was not exposed on the `GrapheneRun` GQL type so
it is added here.

## How I Tested These Changes

Viewed backfill runs page linked by user in shadow dagit with this
applied, confirmed sorting was fixed.

Before:

<img width="270" alt="image"
src="https://github.com/dagster-io/dagster/assets/1531373/46db88a2-6f9e-4f5a-a550-19f831dc2233">

After:

<img width="270" alt="image"
src="https://github.com/dagster-io/dagster/assets/1531373/ad8e002a-1574-4ec7-a098-f9eea8db51e9">