[bug] Fix issue where metadata got discarded in SourceAsset conversion (#22862)
## Summary & Motivation

Resolves: https://github.com/dagster-io/dagster/issues/22789

Fixes issue which I believe was introduced by:
https://github.com/dagster-io/dagster/issues/22165 (didn't dig that
hard)

The core problem is this conversion:
https://sourcegraph.com/github.com/dagster-io/dagster/-/blob/python_modules/dagster/dagster/_core/definitions/external_asset.py?L176

`observable_source_asset` does not set metadata on the
`OutputDefinition` of its internal op. So when we initially convert a
SourceAsset into an ol' fashioned `AssetsDefinition`, then call
`to_source_assets()` on that, we end up losing the metadata field, as we
were trying to pull it off of the `OutputDefinition`. This PR just pulls
it off of the spec instead.

## How I Tested These Changes

Added test fails before, passes now.