[asset differ] add "asset removed" change reason (#22870)
## Summary

Allows the asset graph differ to mark an asset as removed in the case
that an asset is present in the base branch but not the comparison
branch.

In the branch deploymenet case this is currently not used, but is useful
for making the asset differ useful in other use-cases (e.g. tracking
change history on a prod branch).

## Test Plan

New unit test.