[DA] Create any_downstream_conditions AutomationCondition (#22863)
## Summary & Motivation

This creates the "`lazy`" AutomationCondition. I have not fully decided
on a name, but I felt like this was potentially more descriptive of what
it's actually doing.

Handling the weird recursive cases is a bit tricky, hence the odd
structure of the tests. Basically if you have:

A: any_downstream_conditions
B: any_downstream_conditions & foo
C: bar

, A->B->C, and you want to evaluate A, the naive approach would result
in the following infinite loop behavior:

1. convert any_downstream_conditions to (any_downstream_conditions &
foo, bar)
2. this expands to ((any_downstream_conditions & foo, bar) & foo, bar)
3. etc.

The way I avoid this is by not including a condition in the expansion if
it's already part of the "call stack". So in this case, in step 2 we
would just expand `any_downstream_conditions & foo` to ((bar) & foo),
resulting in a full expression of:

(((bar) & foo), bar)

This is certainly a bit duplicative, but it at least feels somewhat
logical and prevents the infinite recursion issue pretty nicely.
Generally this is the sort of danger when you have a mix of regular
`any_downstream_conditions and stuff that combines that with other
filters. Not sure if we would even recommend that pattern, just trying
to get ahead of the flat-out errors

## How I Tested These Changes