[DA] Fix evaluation history UI (#22859)
## Summary & Motivation

Resolves: https://github.com/dagster-io/dagster/issues/22787

The new Declarative Automation framework allows conditions which are
evaluated against parents of a given asset. Those parents do not
necessarily have the same partitions definition as the child.

The current UI was designed with the expectation that all
sub-evaluations would reference the same PartitionsDefinition, and so
results in errors when we (e.g.) try to construct
a`GraphenePartitionedAssetConditionEvaluationNode` from an evaluation
that was actually executed against an unpartitioned asset.

We are working on a larger redesign which should help remove this
assumption from the core UI model, but for now we can prevent these
errors by making it possible to construct a
`...PartitionedEvaluationNode` from an unpartitioned evaluation and
vice-versa.

The end result is a comprehensible UI without errors, although
compromises needed to be made (e.g. an unpartitioned row in the
partitioned UI ends up being represented as having a "partition key" of
the string "None"). This feels like an acceptable state for the short
term.

## How I Tested These Changes