[perf] Remove n^2 logic in asset catalog (#22930)
## Summary & Motivation

I'm not really sure what the original code was trying to do, it seems to
have morphed a bit over the years. I believe just filtering by
prefixPath is all we need to do here so removing all of the extraneous
logic. It's totally possible I'm missing something but from my testing
this works the same.

## How I Tested These Changes


Loaded an asset catalog with 100k assets and saw that loading a folder
view in that catalog didn't cause the page to hang. Also confirmed that
the assets displayed for that folder were correct.