[ui] Use newer virtualizer for logs scrolling table (#22902)
## Summary & Motivation

Use `@tanstack/react-virtual` for the logs scrolling table instead of
`react-virtualized`. If I've done this correctly and haven't overlooked
any features, the implementation of this component is now *much*
simpler.

This was prompted by wanting to add some empty stats for cases where the
filter settings yield no results, which was surprisingly difficult and
buggy with our `react-virtualized` implementation. It should be much
simpler with `@tanstack/react-virtual`.

There are only a couple of `react-virtualized` callsites left in the
codebase.

We'll want to stress test this quite a bit just to make sure it doesn't
introduce unexpected bugs.

## How I Tested These Changes

View an existing run, verify that the virtualized scroll works correctly
in the log table. Verify that:

- Filtering behaves correctly.
- Highlighting for text and timestamp works correctly.
- The loading state is correct.

Kick off a `log_spew` run, verify that logs are appended correctly, and
that the scroll state remains pinned to the bottom. Scroll up manually,
verify that pinning is undone. Scroll back to the bottom, verify that
pinning resumes.