[ui] IA: Settings -> Deployment (#22934)
## Summary & Motivation

For the IA flag, rename "Settings" to "Deployment". This includes
changing `/settings` routes to `/deployment`.

There is a corresponding internal change as well.

## How I Tested These Changes

Inside and outside IA feature flag, verify correct rendering and
behavior of the top nav item and all routes and pages within that
section of the app.