feat(catalog): add `AssetSelection` method to select owners (#23003)
## Summary & Motivation
We should be able to serialize an asset selection for owners, as you can
select for owners in the catalog UI.

We should also do the same for an asset selection for code locations,
but that involves changing how the BaseAssetGraph is constructed. So
separating that into its own change.

## How I Tested These Changes
pytest