feat(dbt): support column lineage for self-referential incremental models (#23024)
## Summary & Motivation
When writing an incremental model, you can use `{{ this }}` as a Jinja
variable to self-reference the current model when generating the next
batch of incremental rows. This made building column lineage fail, as we
did not also pass the schema of the current model to the `sqlglot`
optimization function.

This fixes that.

## How I Tested These Changes
pytest -- updated incremental test would fail if current model's schema
was not passed into `sqlglot` optimization.