Add `target` param to `{Schedule,Sensor}Definition` (#22569)
## Summary & Motivation

Add new `target` param to `{Schedule,Sensor}Definition` and their
associated decorators. An instigator target is an executable entity that
can be targeted by a schedule/sensor. This is
intended as a replacement for the `job`/`job_name` parameters.

This PR builds off downstack work that consolidated `DirectTarget` and
`RepoRelativeTarget` into a single `AutomationTarget` class. The primary
addition here is an `AutomationTarget.from_coercible` class method that
will accept an `AssetsDefinition` or `CoercibleToAssetSelection` and
generate an anonymous asset job from it which is used as the job def for
the `AutomationTarget`.

`AssetsDefinitions` that are passed directly as an automation target,
rather than as `Definitions.assets`, are "hoisted" into the pool of
assets for the `Definitions` object. This means that if asset "foo" is
embedded as an automation target as some sensor, and an unrelated job is
defined by `AssetSelection.assets("foo")`, the job will successfully
resolve foo.

## How I Tested These Changes

New unit tests.