Create exapndedLabel GQL field for AutomationConditionEvaluationNode (#23010)
## Summary & Motivation

As title, this creates an expanded label field which recursively builds
a string to be rendered in the UI. The logic is that an "expandedLabel"
will ditch the user-provided label for that node (if any), and then
render the full expression based on the in-code names of the underlying
conditions.

It adds in a new "name" property to AutomationCondition, which is
necessary because we can't repurpose "description" because that's meant
to be a human-readable description of the precise situations under which
the condition is true (i.e. it also contains information about the
specific parameters of the computation), whereas the name is meant to be
more of a machine / system property. The label needs to be fully
user-customizable and rather than referring to a single node, it
actually refers to the entire subtree underneath that node as well.

By convention, the names of operators are all-caps, and the names of
operands are all lower-case.

In the future, the existing description would likely be rendered in the
UI behind some sort of tooltip, but we can punt on that for now.

## How I Tested These Changes