feat(catalog): add `AssetSelection` stub to select by code location (#23052)
## Summary & Motivation
Add an `AssetSelection` subclass that represents a selection by code
location.

Its selection method isn't implemented yet since we don't expect users
to use this in code. Code location information for a given node in the
asset graph is only available in a remote context, not an in-process
context.

## How I Tested These Changes
pytest