feat(dbt): Add `SdfCliResource` (#23046)
## Summary & Motivation

This PR adds the `SdfCliResource` which is responsible for making calls to the sdf cli binary, and interpreting the results through log / event processing. 

By default, we invoke the  sdf cli with `--log-level=info` that tell dagster when a table has finished materializing etc. We do not extract table schema, or other metadata from structured logs, as this is all available at compile time via the information schema.

## How I Tested These Changes
pytest