[dagster-aws] add PipesGlueClient (#22968)
## Summary & Motivation
Add a new PipesGlueClient to `dagster_aws.pipes`. 

It includes a dummy `PipesGlueMessageReader` class which is currently not usable and is meant to establish an interface (the default MessageReader) for the future. It will become usable once `PipesCloudWatchMessageReader` is implemented. Currently, attempting to use it will raise a `NotImplementedError`. 

A `PipesS3MessageReader` can be used instead right now. 

## How I Tested These Changes

Added a test which uses `moto` and also emulates AWS Glue job execution by running it locally