Fix `useStateWithStorage`'s return type  (#23070)
## Summary & Motivation

The return type lies to you. it tells you that you might get undefined
as an argument to your setter but this is untrue. The argument you
receive is always guaranteed to be type T, and if undefined is not
included in it then you'll never get it.

this addresses pr feedback from
https://github.com/dagster-io/internal/pull/10591/files#r1681728346

## How I Tested These Changes
buildkite