[ui] Asset detail: fix missing materializations error (#23092)
## Summary & Motivation

In the asset detail view, we're encountering an error when reading
`assetMaterializations` for assets that don't exist in the workspace.
Make the code a little more defensive.

## How I Tested These Changes

View an asset that exists in the workspace, verify that the page loads
with no errors. View an asset that no longer exists, verify same.