[DA] Rename *AssetCondition* (#23113)
## Summary & Motivation

We have a set of classes which contain the name "AssetCondition" instead
of "AutomationCondition". This renames those classes.

Note that some of these classes are whitelisted for serdes, and
therefore have "old_storage_names" set. In theory, this is actually not
necessary, as these classes actually did not get serialized in practice
unless someone was using unpublicized APIs, but it doesn't hurt to allow
for this backwards compatibility.

At some point in the future, we could safely remove the
old_storage_names bit, but that can be part of a larger effort of
stripping out obsolete backwards-compatibility code once the Declarative
Automation becomes non-experimental.

## How I Tested These Changes