[ui] Set Fuse options to search entire string (#23142)
## Summary & Motivation

Our current Fuse options won't allow searching to the end of longer
strings, which means that when searching for an asset key element at the
end of a long path might, even an exact string match might be ignored.

To repair this, use `ignoreLocation` since we don't actually care where
in the string we find the match.
https://www.fusejs.io/concepts/scoring-theory.html

I didn't notice any unwanted side effects from this change, but we
should keep an eye out for unexpected behavior.

## How I Tested These Changes

View Catalog as user experiencing this issue. Search for the key
element, verify that the desired asset appears in the search (along with
a number of other good matches that were previously hidden).