Don't use `isEqual` to compare groups in `useAssetDefinitionFilterState` (#23165)
## Summary & Motivation

Don't use `isEqual` to compare groups in `useAssetDefinitionFilterState`
because we're reusing the logic in catalog views and in that context the
filters have a different ` __typename`. I could remove the typename
before passing it to this abstraction but its more performant to just
skip the intermediate objects and just directly compare the attributes
we care about.

## How I Tested These Changes

used this in catalog views 
buildkite.