[ui] Hide jobs with dunder anonymous prefix (#23185)
## Summary & Motivation

@smackesey is introducing a new anonymous job prefix. We should handle
this the same as the existing `__ASSET_JOB` prefix, which is to hide it
aggressively throughout the app.

The simplest thing to do here seems to be to just add it as another
check alongside the existing prefix check.

## How I Tested These Changes

Run dagster dev with a target code location that includes sensors and
schedules with anonymous jobs using the new prefix.

View Jobs list, verify that these jobs don't appear. View Schedules and
Sensors lists, verify same.