[Filters] Memoize filter objects + renderFilterButton (#23230)
## Summary & Motivation

Two things:


1. We're passing the filterButton upwards to the CatalogViewProvider so
that it can use it without having to recreate it from scratch. While
doing that I encountered an infinite loop due to the provider updating
the context which causes `useAssetCatalogFiltering.cloud` to re-render
which causes `useAssetCatalogFiltering.oss` to re-render and it ends up
creating new filter objects due to some aspects of it being unmemoized.
To fix this I memoized things that were causing unnecessary re-renders.

2. Added a `renderButton` return to `useFilters` which can be used to
change the label on the button. I opted to just return a new prop
instead of modifying the existing one to this in order to reduce the
number of changes in this PR. In a future pr we can consolidate or
figure out a more elegant approach.

## How I Tested These Changes

host cloud.