Add AssetPartitionWipeRange class (#22245)
## Summary & Motivation

Add an `AssetPartitionWipeRange` class to complement
`AssetKeyPartitionKey`. This is used upstack in partition wiping PRs.

Like `AssetKeyPartitionKey`, this uses `None` for the `partition_range`
if there is no range-- this is interpreted as "the whole asset".

This is called `AssetPartitionWipeRange` instead of just
`AssetPartitionRange` to scope this abstraction explicitly to partition
wiping for the time being. Later I expect to rename this to just
`AssetPartitionRange` or similar as part of a broader rename of
partition abstractions.

## How I Tested These Changes

Existing test suite.