[dagster-sdf] Enable Upstream Dependencies (#23218)
## Summary & Motivation

This PR upgrades the dagster-sdf minimum version dependency on sdf-cli
to the latest release (0.3.13) which features a new `sdf compile --stage
parse`. This parsing stage produces the results of SDF's dependency
analysis as a parquet table in the information schema directory called
table_deps which features a new `meta` column.

To build on the `meta` column, users are now able to attache arbitrary
metadata to `table.sdf.yml`'s, including most importantly a
`dagster-asset-key`:

```
table:
  name: moms_flower_shop.pub.raw_customers
  meta:
    dagster-asset-key: build_raw_customers
```

This key is used to correlate a source in sdf with it's corresponding
dagster asset. It enables users to specify dagster assets as upstream
dependencies of sdf models materialized by the dagster-sdf integration.

## How I Tested These Changes

See dagster_sdf_tests/test_asset_decorator.py